requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
redis
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
pytest
pytest-asyncio
ruff
email-validator
itsdangerous
mcstatus
Pillow
apscheduler
